requirements.txt
discord==2.3.2
SQLAlchemy==2.0.20
SQLAlchemy-serializer==1.4.1
python-dotenv==1.0.0
aiosqlite==0.19.0
Pillow~=10.0.1